build: build the library in "partial" ivy format
- delivers a library in the "partial" ivy format
- merges a cycle between `NgbDatepicker` and `NgbDatepickerMonth`
- removes unnecessary "ngb:build-ivy" target previously used to catch compilation errors

Closes #4119
