+++ package.json
@@ -11,7 +11,7 @@
     "version": "yarn changelog && git add CHANGELOG.md",
     "postversion": "yarn ngb:build && yarn demo:build",
     "preinstall": "node misc/preinstall.js",
-    "build": "yarn ngb:build-ivy && yarn ngb:build && yarn demo:build",
+    "build": "yarn ngb:build && yarn demo:build",
     "test": "yarn check-format && yarn ngb:lint && yarn ngb:test && yarn schematics:test",
     "tdd": "yarn ngb:tdd",
     "e2e": "yarn e2e-app:lint && yarn ngb:e2e && yarn demo:e2e",
@@ -35,7 +35,6 @@
     "ngb:tdd": "ng test ng-bootstrap --source-map false",
     "ngb:e2e": "concurrently --success first --kill-others \"yarn e2e-app:serve -c playwright\" \"yarn ngb:e2e-playwright && yarn ngb:e2e-playwright -b firefox && yarn ngb:e2e-playwright -b webkit\"",
     "ngb:e2e-playwright": "yarn playwright",
-    "ngb:build-ivy": "ng build ng-bootstrap",
     "ngb:build": "ng build ng-bootstrap -c production && yarn schematics:build && yarn ngb:static",
     "demo:serve": "ng serve demo --host 0.0.0.0",
     "demo:docs": "ts-node --project misc/tsconfig.json misc/generate-docs.ts",
