+++ src/datepicker/datepicker.module.ts
@@ -2,7 +2,7 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {NgbDatepicker, NgbDatepickerContent} from './datepicker';
-import {NgbDatepickerMonth} from './datepicker-month';
+import {NgbDatepickerMonth} from './datepicker';
 import {NgbDatepickerNavigation} from './datepicker-navigation';
 import {NgbInputDatepicker} from './datepicker-input';
 import {NgbDatepickerDayView} from './datepicker-day-view';
@@ -17,7 +17,7 @@ export {NgbCalendarPersian} from './jalali/ngb-calendar-persian';
 export {NgbCalendarHebrew} from './hebrew/ngb-calendar-hebrew';
 export {NgbDatepickerI18nHebrew} from './hebrew/datepicker-i18n-hebrew';
 export {NgbCalendarBuddhist} from './buddhist/ngb-calendar-buddhist';
-export {NgbDatepickerMonth} from './datepicker-month';
+export {NgbDatepickerMonth} from './datepicker';
 export {NgbDatepickerDayView} from './datepicker-day-view';
 export {NgbDatepickerNavigation} from './datepicker-navigation';
 export {NgbDatepickerNavigationSelect} from './datepicker-navigation-select';
